Big commit here... Args working
Arguments are working and I have tested pretty thoroughly for any
memory leaks. In addition, any number of args (fewer than MAXARGC) may
be passed now to a child process. Actual execution now uses unix forks
like it should. I used the execvpe system call; I'm not sure if this is
actually allowed on the project though, or will be in all unix systems.
Need to make sure this will work fine on the test machine!